# Conflicts:
#	src/main/java/com/mp/karental/controller/BookingController.java
#	src/main/java/com/mp/karental/repository/BookingRepository.java
#	src/main/java/com/mp/karental/service/BookingService.java
#	src/main/java/com/mp/karental/service/CarService.java
#	src/test/java/com/mp/karental/controller/BookingControllerTest.java
#	src/test/java/com/mp/karental/controller/CarControllerTest.java
#	src/test/java/com/mp/karental/service/BookingServiceTest.java
#	src/test/java/com/mp/karental/service/CarServiceTest.java
#	src/test/java/com/mp/karental/validation/validator/BookingTimeValidatorTest.java
#	src/test/java/com/mp/karental/validation/validator/StatusEditValidatorTest.java

+++ src/main/java/com/mp/karental/dto/request/EditCarRequest.java
@@ -30,10 +30,10 @@ public class EditCarRequest {
     float fuelConsumption;
     @Min(value = 0, message = "INVALID_VALUE_MIN")
     @RequiredField(fieldName = "Car's base price")
-    int basePrice;
+    long basePrice;
     @Min(value = 0, message = "INVALID_VALUE_MIN")
     @RequiredField(fieldName = "Deposit value")
-    int deposit;
+    long deposit;
 
     @RequiredField(fieldName = "Address")
     @ValidAddress(message = "INVALID_ADDRESS")
