+++ src/main/java/com/mp/karental/dto/response/CarResponse.java
@@ -31,8 +31,8 @@ public class CarResponse {
     int productionYear;
     float mileage;
     float fuelConsumption;
-    int basePrice;
-    int deposit;
+    long basePrice;
+    long deposit;
     String address;
     String description;
     String additionalFunction;
